refine models actions to use groq sdk + get completion
