+++ eval-dash/.gitignore
@@ -19,6 +19,7 @@
 .DS_Store
 *.pem
 migrations
+*_backup*
 
 # debug
 npm-debug.log*
