+++ eval-dash/package.json
@@ -30,6 +30,7 @@
     "@types/node": "22.10.1",
     "@types/react": "19.0.0",
     "@types/react-dom": "19.0.0",
-    "drizzle-kit": "^0.30.1"
+    "drizzle-kit": "^0.30.1",
+    "groq-sdk": "^0.9.1"
   }
 }
